#  Use less stack for HttpResponseHeaders.CopyToFast