… src/Servers/Kestrel/Core/src/Internal/Http/HttpRequestHeaders.cs
… src/Servers/Kestrel/Core/src/Internal/Http/HttpResponseHeaders.cs
… src/Servers/Kestrel/Core/src/Internal/Http/HttpResponseTrailers.cs
… src/Servers/Kestrel/perf/Kestrel.Performance/ResponseHeadersWritingBenchmark.cs
… src/Servers/Kestrel/tools/CodeGenerator/KnownHeaders.cs