The Jit seems unhappy to optimize this method due to the number of locals and tmps this method creates (585 locals/tmps; including all the new `ReadOnlySpan<byte>`s).

This change slims the method by 9638 bytes of asm from 13112 bytes to 3474 bytes; and reduces the stack clearing from 416 bytes to 8 bytes,

```diff
  Method |                       Type |     Mean |         Op/s |
 ------- |--------------------------- |---------:|-------------:|
- Output |                 LiveAspNet | 587.5 ns |  1,702,161.8 |
+ Output |                 LiveAspNet | 462.2 ns |  2,163,172.9 |
- Output |           PlaintextChunked | 139.8 ns |  7,151,853.6 |
+ Output |           PlaintextChunked |  84.3 ns | 11,850,576.2 |
- Output | PlaintextChunkedWithCookie | 276.3 ns |  3,618,669.2 |
+ Output | PlaintextChunkedWithCookie | 179.5 ns |  5,570,065.5 |
- Output |        PlaintextWithCookie | 289.5 ns |  3,454,397.7 |
+ Output |        PlaintextWithCookie | 169.2 ns |  5,908,899.3 |
- Output |       TechEmpowerPlaintext | 175.4 ns |  5,700,681.7 |
+ Output |       TechEmpowerPlaintext |  86.1 ns | 11,607,037.7 |
```

It does this by using a single output point for the headers (other than the unusual ones Raw and ContentLength) where a single ReadOnlySpan is created. To achieve this it is a little convoluted using goto and switch... 

Before
```asm
; Lcl frame size = 1704

G_M45189_IG01:
       4157                 push     r15
       4156                 push     r14
       4155                 push     r13
       4154                 push     r12
       57                   push     rdi
       56                   push     rsi
       55                   push     rbp
       53                   push     rbx
       4881ECA8060000       sub      rsp, 0x6A8
       C5F877               vzeroupper 
       488BF1               mov      rsi, rcx
       488D7C2428           lea      rdi, [rsp+28H]
       B9A0010000           mov      ecx, 416           ; zero 416 bytes of stack
       33C0                 xor      rax, rax
       F3AB                 rep stosd 
       
       ; ...

; Total bytes of code 13112, prolog size 42 for method HttpResponseHeaders:CopyToFast(byref):this
```

```asm
; Lcl frame size = 88

G_M44993_IG01:
       4157                 push     r15
       4156                 push     r14
       4155                 push     r13
       4154                 push     r12
       57                   push     rdi
       56                   push     rsi
       55                   push     rbp
       53                   push     rbx
       4883EC58             sub      rsp, 88
       488BF1               mov      rsi, rcx
       488D7C2430           lea      rdi, [rsp+30H]
       B908000000           mov      ecx, 8          ; Zero 8 bytes of stack
       33C0                 xor      rax, rax
       F3AB                 rep stosd 
       
       ; ...
       
; Total bytes of code 3474, prolog size 42 for method HttpResponseHeaders:CopyToFast(byref):this
```

/cc @AndyAyersMS