@Tratcher 2019-02-19
@jkotalik header ordering is not meaningful, especially if you avoid re-ordering multiple instances of the same header.