@AndyAyersMS 2019-02-19
Keeping the number of jit temps under 512 is advisable in perf-sensitive code.

512 is the maximum number of things the jit will model for liveness (which is a "dense" data flow problem that requires using bitvectors, so lots of jit time & memory).

See dotnet/coreclr#13280 for some context.

When you go beyond this number, the jit marks 512 temps as "tracked" temps and the remainder as "untracked" temps. It chooses the set of tracked temps greedily, using block-weighed aggregate appearances as a priority function. 

Untracked GC ref temps prolog zeroing. In a method like this where average path length is fairly small (even though method itself is quite large) the prolog costs can predominate. There are a bunch of potential mitigation strategies in the JIT for this, but nothing that is being actively worked on right now.

I wonder though  -- if the JIT sees we're in the untracked regime, perhaps it should somewhat prioritize tracking GC refs over non-GC temps... hmm (alternatively: consider an untracked GC ref incurs an implicit extra "prolog" weight from the zeroing that we don't account for -- perhaps we should make that explicit). The goal should really be to minimize the overall post-sort cost, and we don't quite do that right now. Let me play around with this idea for a bit.