@benaadams 2019-02-19
Changed to maintain ContentLength's header position