@AndyAyersMS 2019-02-19
At least in the before example, better temp sorting by the JIT would not help reduce prolog cost. Despite there being 585 temps, all the ref type temps, and all but one of the byref type temps end up getting tracked. EG the `T197` below is the tracking ID for the temp.
```
;  V581 tmp472      [V581,T197] (  2,  8   )   byref  ->  rcx         "argument with side effect"
...
;  V353 tmp244      [V353    ] (  4,  1.75)   byref  ->  [rsp+0x1D0]   must-init pinned "Inline stloc first use temp"
```
`V353` is the only untracked scalar GC ref.

The bulk of of the prolog zeroing thus comes from GC refs within structs. There are 118 struct temps, all marked as `must-init` (the result of `.init locals` on the method), and this accounts for the large amount of zeroing. This can be removed by the current jit only if the struct is promoted, but it doesn't look like any of these are promoted., they are all marked "do not enregister." Probably worth understanding why that happens.