@benaadams 2019-02-20
A fair boost on the `ResponseHeadersWritingBenchmark`
```diff
  Method |                       Type |     Mean |         Op/s |
 ------- |--------------------------- |---------:|-------------:|
- Output |                 LiveAspNet | 587.5 ns |  1,702,161.8 |
+ Output |                 LiveAspNet | 511.5 ns |  1,955,163.5 |
- Output |           PlaintextChunked | 139.8 ns |  7,151,853.6 |
+ Output |           PlaintextChunked |  86.4 ns | 11,574,316.8 |
- Output | PlaintextChunkedWithCookie | 276.3 ns |  3,618,669.2 |
+ Output | PlaintextChunkedWithCookie | 194.3 ns |  5,148,107.5 |
- Output |        PlaintextWithCookie | 289.5 ns |  3,454,397.7 |
+ Output |        PlaintextWithCookie | 181.2 ns |  5,520,292.8 |
- Output |       TechEmpowerPlaintext | 175.4 ns |  5,700,681.7 |
+ Output |       TechEmpowerPlaintext | 101.1 ns |  9,891,448.7 |
```