@AndyAyersMS 2019-02-20
And as for why there are so many unpromoted structs in the before version -- perhaps no great mystery: we run into the too many temps throttling in the inliner, after 328 successful inlines. 

Similar to the kinds of things I'm tracking in dotnet/coreclr#22240.
```
  [328 IL=1167 TR=003047 06000071] [profitable inline] StringValues:get_Count():int:this
  [0 IL=1213 TR=003105 06001527] [FAILED: too many locals] ReadOnlySpan`1:.ctor(ref,int,int):this
  [0 IL=1218 TR=003111 06000017] [FAILED: too many locals] BufferWriter`1:Write(struct):this
  [0 IL=1226 TR=003120 060006E2] [FAILED: too many locals] PipelineExtensions:WriteAsciiNoValidation(byref,ref)
  [0 IL=1278 TR=002935 06000071] [FAILED: too many locals] StringValues:get_Count():int:this
  ... and many more ....
  [0 IL=4338 TR=000426 06001527] [FAILED: too many locals] ReadOnlySpan`1:.ctor(ref,int,int):this
  [0 IL=4343 TR=000432 06000017] [FAILED: too many locals] BufferWriter`1:Write(struct):this
  [0 IL=4351 TR=000441 060006E2] [FAILED: too many locals] PipelineExtensions:WriteAsciiNoValidation(byref,ref)
```
Probably worth running TechEmpower scenarios with JIT inline ETL enabled to see if this is a common failure reason, and if so, where we might want to reduce method complexity.

Am going to temporarily remove this check and let the inliner run free, just to see what happens:

https://github.com/dotnet/coreclr/blob/e2081d0e67a1d7fedaf6303f576acef316c7bd66/src/jit/compiler.hpp#L1531-L1535