@benaadams 2019-02-20
Plaintext inlines https://gist.github.com/benaadams/1e08f4fe89b3a07a64a915d388a63336