@AndyAyersMS 2019-02-20
Inliner can handle it, but not the rest of the jit. Way more temps, lots of untracked, etc.
```
Top method regressions by size (bytes):
       17684 (134.87% of base) : Microsoft.AspNetCore.Server.Kestrel.Core.dasm - HttpResponseHeaders:CopyToFast(byref):this

;  V1249 tmp1140    [V1249,T506] (  2,  8   )   byref  ->  rcx         "argument with side effect"
;  V1250 tmp1141    [V1250,T507] (  2,  8   )   byref  ->  rcx         "argument with side effect"
;  V1251 tmp1142    [V1251,T508] (  2,  8   )   byref  ->  rcx         "argument with side effect"
;  V1252 tmp1143    [V1252,T509] (  2,  8   )   byref  ->  rcx         "argument with side effect"
;  V1253 cse0       [V1253,T01] (207,364.50)   byref  ->  [rsp+0x28]   "ValNumCSE"
;
; Lcl frame size = 7024

G_M45944_IG01:
       push     r15
       push     r14
       push     r12
       push     rdi
       push     rsi
       push     rbp
       push     rbx
       test     qword ptr [rsp-1000H], rax
       sub      rsp, 0x1B70
       vzeroupper 
       mov      rsi, rcx
       lea      rdi, [rsp+30H]
       mov      ecx, 0x6D0
```
That looks like the only method in the core assembly that hits this limit.

Simplifying it is the right thing to do. 

We may get the jit to the point where it can handle things like the before case better someday, but that day is a ways off.