@benaadams 2019-02-20
Cleaned up the other switch jump tables to use named labels rather than just numbers so its easier to compare the label name to what its outputting