@benaadams 2019-02-22
Changed to loop; perf is very nice (also has ROS opt etc, so isn't like for like on before loop)
```csharp
 Method |                       Type |      Mean |         Op/s |
------- |--------------------------- |----------:|-------------:|
 Output |                 LiveAspNet | 462.28 ns |  2,163,172.9 |
 Output |           PlaintextChunked |  84.38 ns | 11,850,576.2 |
 Output | PlaintextChunkedWithCookie | 179.53 ns |  5,570,065.5 |
 Output |        PlaintextWithCookie | 169.24 ns |  5,908,899.3 |
 Output |       TechEmpowerPlaintext |  86.15 ns | 11,607,037.7 |
```