@pakrym 2019-02-22
Any changes in public APIs? Rebase and run `.\eng\scripts\GenerateReferenceAssemblies.ps1`