@benaadams 2019-02-22
CI issues https://github.com/aspnet/AspNetCore/issues/7839 and https://github.com/aspnet/AspNetCore/issues/7867